Final version 42 (#185)
* all builtins normed expect cd

* fixed the prompt printing

* prompt in green

* argv.c bash_pattern.c bash_sym.c normed

* norminette -> command.c debugger.c env.c env_access.c environ.c

* env_node.c env_singleton.c exex_bin.c execscheme.c normed

* comment line unitest in environ.cpp

* executor.c normed

* executor_dispatch executor_scheme_pipe and redirect normed

* in the norming journey

* norming the rest

* norm

* norming

* fixing what I broke while norminetting

* fixing problems

* continuing norm

* almost normed all .c files

* all .c normed

* norming .h files

* bug fixed in parser.c

* executor.c: changed back order of execution

* src/process_table.c ok

* src/process_table.c finally good

* bash_pattern.c ok

* token.c -> ok

* file_system -> ok

* execscheme done

* env_access -> ok

* process -> ok

* env -> ok

* debugger -> used extern for global variables seem to work

* env_node and executor -> ok

* environ and prompt -> ok

* lexer -> ok

* env_singleton -> ok

* parser -> ok

* journal done, all .c files normed finally

* .h files almost normed

* Bash test 42 (#188)

* bash_test.sh / bash_test_continuous.sh: simple bash tests

* bash_test.sh: added tests for builtins

* Final version 42 stanedits (#191)

* bash_test.sh / bash_test_continuous.sh: simple bash tests

* bash_test.sh: added tests for builtins

* bumped libgnl & libvector with build options and arraybound fixes

* bumped libgnl: named test

* bash_test.sh: reworked for doing multiple tests

* minimake && Cmakelists : reworked for release/debug production

* builtin_pwd: fixed memory leak

* shellsignals.c: static is now volatile, removed unncessary functions

* process_queue.c: volatile

* process.c: added return statement

* journal_continue.c: set default value

* process_table.c: cleanup for global

* CMakeLists.txt: disabled all optimization

* argvc.c: added NULL ptr to end of arglist

* gitignore: added vcore for valgrind

* argvc.cpp: added simple test for argv

* bash_test.sh: minor updates

* executors: progressing

* shellsignals.c: added dbg() statement to sighandler

* shellcore.c: added dbg() statements

* prompt_singleton.c: deinit now reset glob ptr

* process_queue.c: register for signal unconditionally

* execscheme.c: optimization

* process_queue.c: exits on unknown signal

* process_continue.c: now using abolished signal(), since 42 only allows that..

* parser_continue2.c: skipping space for builtin checking

* parser: added debugging statements

* executor: runtime improvements

* debugger.h: better offset of output

* env_access.h: added env_set_lasterror()

* env_singleton.c/h: added env_set_s_line()

* execscheme.h: added pid_t to execscheme struct

* proccess, various -> reworked pid logic into execscheme

* parser, various -> bugfixes, better offloading to executor

* apps, various -> api changed in parse() applied

* ft_unistd.h: added drop_pipe()

* executor, various -> completely reworked execution logic

* journal.c: minor fixes:

* builtins/exec_bin/assignment -> reworked for new execution logic

* env, various -> bugfixes and added env_set_lasterror()

* pipe.c: added drop_pipe()

* environ.c: added V_REPLACE functionality to environ_add

* preparser.c: added preparsing logic

* prompt.c: simplified prompt last_error handling

* shellcore.c: using new execution logic with 'preparsing'

* shellsignals.c: removed p_tab logic, for some reason it still works

* bash_test.sh: now uses valgrind as well

* CMakeLists.txt: granular compilation options added

* bash_test.sh: added timeout for valgrind check

* CMakeLists.txt: removed optimization flags since this seems be be borking get_next_line or prompt

+++ src/process.c
@@ -12,53 +12,8 @@
 #include "vector.h"
 #include "process.h"
 
-pid_t	*p_allocate_pid(pid_t pid)
+static int	signal_status(t_waitcond cond, int status)
 {
-	return ((pid_t *)ft_memdup(&pid, sizeof(pid_t)));
-}
-
-/* man 2 sigaction */
-int		p_register_signalhandler(int sig, void (*sighandler)(int sig))
-{
-	struct sigaction sa;
-
-	sa.sa_handler = sighandler;
-	sigemptyset(&sa.sa_mask);
-	sa.sa_flags = SA_RESTART;
-	if (sigaction(sig, &sa, NULL) == -1)
-	{
-		dbg("couldn't register handler for signal %s, errno: %s\n", strsignal(sig), strerror(errno));
-		return (-1);
-	}
-	return (0);
-}
-
-int		p_pause()
-{
-	pause();
-	if (errno == EINTR)
-		dbg("pid %i caught and handled signal, continuing..\n", getpid());
-	return (0);
-}
-
-static t_bool	__correct_signal(t_waitcond cond, int status)
-{
-	/* doesn't follow open/close principle, but switch case not allowed */
-
-	if (cond == W_EXITED)
-		return (WIFEXITED(status));
-	else if (cond == W_TERMINATED)
-		return (WIFSIGNALED(status));
-	else if (cond == W_ABORTED)
-		return (WIFSIGNALED(status) && WCOREDUMP(status));
-	else
-		assert(!"UNDEFINED WAIT CONDITION");
-}
-
-static int		__return_signal_status(t_waitcond cond, int status)
-{
-	/* doesn't follow open/close principle, but switch case not allowed */
-
 	if (cond == W_EXITED)
 		return (WEXITSTATUS(status));
 	else if (cond == W_TERMINATED)
@@ -67,34 +22,78 @@ static int		__return_signal_status(t_waitcond cond, int status)
 		return (WCOREDUMP(status));
 	else
 		assert(!"UNDEFINED WAIT CONDITION");
+	return (FALSE);
 }
 
-/* man 2 waipid */
-int		p_waitpid(pid_t pid, t_waitcond cond)
+static void	dump_signal_status(int status, pid_t pid)
+{
+	if (WIFEXITED(status))
+		dbg("Pid %i exited with status %i!\n", pid, WEXITSTATUS(status));
+	else if (WIFSIGNALED(status))
+		dbg("Pid %i was terminated with status %i!\n", pid, WTERMSIG(status));
+	else if (WIFSIGNALED(status) && WCOREDUMP(status))
+		dbg("Pid %i was aborted with status %i!\n", pid, WCOREDUMP(status));
+	else
+		dbg("Pid %i has illegal status!\n", pid);
+}
+
+int	p_wait(pid_t *pid, t_waitcond cond)
 {
 	int	status;
 
 	if (cond == W_EXITED)
-		dbg("waiting for pid %i to exit\n", pid);
+		dbg("Waiting for pid %i to exit\n", pid);
 
-	while (waitpid(pid, &status, 0) <= 0 && !__correct_signal(cond, status))
-	;;
-
-	if (cond == WEXITED)
-		dbg("pid %i exited with status %i!\n", pid, WEXITSTATUS(pid));
-
-	return (__return_signal_status(cond, status));
+	waitpid(*pid, &status, 0);
+	dump_signal_status(status, *pid);
+	if (signal_status(cond, status) == -1)
+		*pid = -1;
+	else
+		*pid = 0;
+	return (signal_status(cond, status));
 }
 
-int		p_signal(pid_t pid, int signal)
+int	p_signal(pid_t pid, int signal)
 {
 	int	error;
 
+	if (kill(pid, 0) == -1)
+	{
+		dbg("Not allowed to send signal %s to pid %i, errno: %s\n",
+			strsignal(signal), pid, strerror(errno));
+		return (-1);
+	}
 	error = kill(pid, signal);
 	if (error == -1)
 	{
-		dbg("couldn't send signal %s to pid %i, errno: %s\n", strsignal(signal), pid, strerror(errno));
+		dbg("couldn't send signal %s to pid %i, errno: %s\n",
+			strsignal(signal), pid, strerror(errno));
+	}
+	else
+	{
+		dbg("Succesfully sent signal %s to pid %i (from pid: %i)\n",
+			strsignal(signal), pid, getpid());
 	}
 	return (error);
 }
 
+pid_t	*p_allocate_pid(pid_t pid)
+{
+	return ((pid_t *)ft_memdup(&pid, sizeof(pid_t)));
+}
+
+int	p_register_signalhandler(int sig, void (*sighandler)(int sig))
+{
+	if (signal(sig, sighandler) == SIG_ERR)
+	{
+		dbg("Failed to register handler for signum: %i, with errno: %s\n",
+			sig, strerror(errno));
+		return (-1);
+	}
+	else
+	{
+		dbg("Succesfully registered handler for signum: %i for pid %i\n",
+			sig, getpid());
+	}
+	return (0);
+}
